Add error when saving file to install directory
